chore(test): move job attachments e2e tests to separate file
